Fix reading from url and local storage + Move parsing of experiment config to module with same name
