+++ apps/class-solid/src/lib/experiment_config.ts
@@ -1,5 +1,8 @@
 import type { Config } from "@classmodel/class/config";
 import type { PartialConfig } from "@classmodel/class/config_utils";
+import { ValidationError, ajv } from "@classmodel/class/validate";
+import type { DefinedError, JSONSchemaType } from "ajv";
+import { findPresetByName } from "./presets";
 
 /**
  * An experiment configuration is a combination of a reference configuration and a set of permutation configurations.
@@ -17,3 +20,61 @@ export interface ExperimentConfig<C = Config> {
  * Parameters in reference which are same as in preset are absent.
  */
 export type PartialExperimentConfig = ExperimentConfig<PartialConfig>;
+
+/**
+ * JSON schema of experiment config without checking the configs, thats done by the presets[x].parse()
+ */
+const jsonSchemaOfExperimentConfigBase = {
+  type: "object",
+  properties: {
+    preset: { type: "string", default: "Default" },
+    reference: {
+      type: "object",
+      additionalProperties: true,
+    },
+    permutations: {
+      type: "array",
+      items: {
+        type: "object",
+        additionalProperties: true,
+      },
+      default: [],
+    },
+  },
+  required: ["preset", "reference", "permutations"],
+} as unknown as JSONSchemaType<ExperimentConfig<object>>;
+
+const validateExperimentConfigBase = ajv.compile(
+  jsonSchemaOfExperimentConfigBase,
+);
+
+function parseExperimentConfigBase(input: unknown): ExperimentConfig<object> {
+  // TODO make preset aware, aka when validating use defaults from preset
+  if (!validateExperimentConfigBase(input)) {
+    throw new ValidationError(
+      validateExperimentConfigBase.errors as DefinedError[],
+    );
+  }
+  return input;
+}
+
+/** Parse unknown input into a Experiment configuration
+ *
+ * The input can be partial, i.e. only contain the fields that are different from the preset or reference.
+ *
+ * @param input - The input to be parsed.
+ * @returns The validated input as a Experiment configuration object.
+ * @throws {ValidationError} If the input is not valid according to the validation rules.
+ */
+export function parseExperimentConfig(input: unknown): ExperimentConfig {
+  const base = parseExperimentConfigBase(input);
+  const preset = findPresetByName(base.preset);
+  const reference = preset.parse(base.reference);
+  // TODO a reference paramaeer not overwritten in a permutation should be in permutation.
+  const permutations = base.permutations.map((perm) => preset.parse(perm));
+  return {
+    reference,
+    preset: base.preset,
+    permutations,
+  };
+}
